Implemented path cost function and modified related methods to adapt with the new return type.

+++ src/code/Search/Heuristic.java
@@ -3,5 +3,5 @@ package code.Search;
 import code.Problem.State;
 
 public interface Heuristic {
-    int calculate(State s);
+    double calculate(State s);
 }
